requirements.txt
playwright==1.44.0
beautifulsoup4==4.12.3
httpx==0.27.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
alembic==1.13.1
fastapi==0.111.0
uvicorn==0.30.1
apscheduler==3.10.4
python-dotenv==1.0.1
geopy==2.4.1
pydantic==2.7.1
pydantic-settings==2.3.0
structlog==24.2.0
tenacity==8.3.0
pandas==2.2.2
numpy==1.26.4
pytest==8.2.2
pytest-asyncio==0.23.7
groq==0.8.0
